requirements.txt
pymongo
dnspython
python-dotenv
resend
